django/requirements.txt
asgiref~=3.2
jsonfield
sqlparse
certifi
decorator
Django
django-cors-headers
gunicorn
hexbytes
oauthlib
Pillow
ptyprocess
python-dateutil
pytz
wcwidth
websockets
django-environ
psycopg2-binary
PyJWT
djangorestframework
djangorestframework_simplejwt
openpyxl
requests
tzdata
boto3
coverage